aps/rag-agent/requirements.txt
# Core LangChain
langchain==0.3.13
langchain-community==0.3.13
langchain-google-genai==2.0.8

# Vector Database (substituído ChromaDB por FAISS)
faiss-cpu==1.9.0.post1

# Document Processing
pypdf==3.17.4
python-dotenv==1.0.0

# Web scraping (optional)
beautifulsoup4==4.12.2
requests==2.31.0

# Google Gemini
google-generativeai==0.8.3

# Utilities
tiktoken==0.5.2
